chore: refactor MCP server to use AI SDK tools
Replace custom tool implementations with AI SDK tools to reduce code duplication.
Addresses issue #40 by consolidating all MCP tool logic into the AI SDK.

- Use createWriteActionTool, createCheckConnectionTool, createCreateInstallationTool, createCheckInstallationTool, createStartOAuthTool, and createSendRequestTool from @amp-labs/ai/mcp
- Remove duplicate tool implementations (connection.ts, oauth.ts, request.ts, write.ts)
- Add createSendReadRequestTool for backward compatibility
- Maintain same API surface for existing MCP clients
